# api/openapi.yaml
openapi: 3.1.0
info:
  title: go-service-main API
  version: 0.4.0
  description: Single-tenant administration API compatible with vue3-element-admin.
security:
  - bearerAuth: []
paths:
...
  /api/v1/roles:
    get:
      operationId: listRoles
      parameters:
        - { in: query, name: pageNum, schema: { type: integer, minimum: 1 } }
        - { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 200 } }
        - { in: query, name: keywords, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    post:
      operationId: createRole
      requestBody: { $ref: '#/components/requestBodies/RoleForm' }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
...
  /api/v1/roles/{roleId}/form:
    parameters:
      - $ref: '#/components/parameters/RoleID'
    get:
      operationId: getRoleForm
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/roles/{roleId}:
    parameters:
      - $ref: '#/components/parameters/RoleID'
    put:
      operationId: updateRole
      requestBody: { $ref: '#/components/requestBodies/RoleForm' }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    delete:
      operationId: deleteRoles
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/roles/{roleId}/menu-ids:
    parameters:
      - $ref: '#/components/parameters/RoleID'
    get:
      operationId: getRoleMenuIDs
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/roles/{roleId}/menus:
    parameters:
      - $ref: '#/components/parameters/RoleID'
    put:
      operationId: updateRoleMenus
      requestBody: { $ref: '#/components/requestBodies/IDList' }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/roles/{roleId}/dept-ids:
    parameters:
      - $ref: '#/components/parameters/RoleID'
    get:
      operationId: getRoleDepartmentIDs
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/roles/{roleId}/depts:
    parameters:
      - $ref: '#/components/parameters/RoleID'
    put:
      operationId: updateRoleDepartments
      requestBody: { $ref: '#/components/requestBodies/IDList' }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/menus:
    get:
      operationId: listMenus
      parameters:
        - { in: query, name: keywords, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    post:
      operationId: createMenu
      requestBody: { $ref: '#/components/requestBodies/MenuForm' }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/menus/options:
    get:
      operationId: getMenuOptions
      parameters:
        - { in: query, name: onlyParent, schema: { type: boolean } }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
...
  /api/v1/menus/{menuId}/form:
    parameters:
      - $ref: '#/components/parameters/MenuID'
    get:
      operationId: getMenuForm
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/menus/{menuId}:
    parameters:
      - $ref: '#/components/parameters/MenuID'
    put:
      operationId: updateMenu
      requestBody: { $ref: '#/components/requestBodies/MenuForm' }
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    delete:
      operationId: deleteMenu
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
...
  /api/v1/configs/refresh/{key}:
    parameters:
      - { in: path, name: key, required: true, schema: { type: string } }
    post:
      operationId: refreshSystemConfigCacheByKey
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/configs/key/{key}:
    parameters:
      - { in: path, name: key, required: true, schema: { type: string } }
    get:
      operationId: getSystemConfigByKey
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/configs/{configId}/form:
    parameters:
      - { in: path, name: configId, required: true, schema: { type: string } }
    get:
      operationId: getSystemConfigForm
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
  /api/v1/configs/{configId}:
    parameters:
      - { in: path, name: configId, required: true, schema: { type: string } }
    get:
      operationId: getSystemConfig
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    put:
      operationId: updateSystemConfig
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
    delete:
      operationId: deleteSystemConfigs
      description: configId may contain one ID or comma-separated IDs.
      responses:
        '200': { $ref: '#/components/responses/AdminAPISuccess' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    RoleID:
      in: path
      name: roleId
      required: true
      schema: { type: string }
    MenuID:
      in: path
      name: menuId
      required: true
      schema: { type: string }
...
  requestBodies:
...
    RoleForm:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/RoleForm' }
    MenuForm:
      required: true
      content:
        application/json:
          schema: { $ref: '#/components/schemas/MenuForm' }
...
    IDList:
      required: true
      content:
        application/json:
          schema:
            type: array
            items: { type: integer, format: int64, minimum: 1 }
  responses:
    AdminAPISuccess:
      description: Successful administration API response.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/AdminAPIResult' }
...
  schemas:
...
    RoleForm:
      type: object
      required: [name, code, status, dataScope]
      properties:
        name: { type: string }
        code: { type: string }
        sort: { type: integer }
        status: { type: integer, enum: [0, 1] }
        dataScope: { type: integer, minimum: 1, maximum: 5 }
        deptIds:
          type: array
          items: { type: string }
    MenuForm:
      type: object
      required: [name, type, visible]
      properties:
        parentId: { type: string }
        name: { type: string }
        type: { type: string, enum: [C, M, E, B] }
        routeName: { type: string }
        routePath: { type: string }
        component: { type: string }
        externalUrl: { type: string }
        perm: { type: string }
        alwaysShow: { oneOf: [{ type: boolean }, { type: integer, enum: [0, 1] }] }
        keepAlive: { oneOf: [{ type: boolean }, { type: integer, enum: [0, 1] }] }
        visible: { type: integer, enum: [0, 1] }
        sort: { type: integer }
        icon: { type: string }
        redirect: { type: string }
        params:
          type: array
          items:
            type: object
            properties:
              key: { type: string }
              value: { type: string }
...
    AdminAPIResult:
      type: object
      required: [code, msg, data]
      properties:
        code: { type: string }
        msg: { type: string }
        data: {}
        requestId: { type: string }